Move ETA model and state into their own module
